Finalizei o crud com o paradigma de programação POO e migrei o banco de dados pro postgreSQL

+++ sistema.py
@@ -28,7 +28,7 @@ print(resultado)
 
 # para atualizar os dados no banco 
 atualizar_tabela = input("Deseja fazer alguma alteração? s/n: ")
-while atualizar_tabela == "s":
+if atualizar_tabela == "s":
     maquina_id = input("Digite o ID da máquina que deseja atualizar: ")
     atualizar_usuario = input("Digite o nome do usuario: ")
     atualizar_patri = input("Digite o novo patrimonio: ")
@@ -38,24 +38,22 @@ while atualizar_tabela == "s":
     conexao.commit()
     print("Máquina atualizada com sucesso!")
     if atualizar_tabela != "s" and atualizar_tabela == "n":
-        break
-    break
+        return 
    
 # para deletar as máquinas
 delete_maquina = input("Deseja deletar alguma máquina?")
-while delete_maquina == "s":
+if delete_maquina == "s":
     deletar_maquina = input("Digite o ID para deletar a máquina: ")
     cursor.execute('DELETE FROM maquinas WHERE id = (%s)', (deletar_maquina,))
     conexao.commit()
     cursor.close()
     print("Máquina deletada com sucesso!")
     if delete_maquina != "s" and deletar_maquina == "n":
-        break
-    break
+        return 
 
 verificacao = input("Você quer adicionar mais máquinas? s/n?: ")
 
-while verificacao == "s":
+if verificacao == "s":
     maquina_usuario = input("digite seu RE:")
     maquina_patri = input("digite o patrimônio:")
     maquina_hostname = input("digite o HOSTNAME: ")
@@ -66,7 +64,6 @@ while verificacao == "s":
     verificacao = input("Você quer adicionar mais máquinas? s/n?: ")
 
     if verificacao == "n":
-        break
+        return
     elif verificacao != "s" and verificacao != "n":
         input("ERRO: digite s/n: ")
-    break
